+++ types/CHANGELOG.md
@@ -29,11 +29,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Input::{is_utxo, as_utxo, is_treasury, as_treasury}`;
 - Derive `Copy` on `UtxoInput`;
 - `Feature::{is_sender, as_sender, is_issuer, as_issuer, is_metadata, as_metadata, is_tag, as_tag}`;
+- `participation` and `indexer` plugin API types from iota-client;
 
 ### Changed
 
 - Updated dependencies;
 - `OutputMetadataResponse` renamed to `OutputMetadataDto` and moved from `api` to `block`;
+- Moved core API types into `core`;
 
 ### Removed